+++ js/changing-size.js
@@ -1,19 +1,14 @@
-const Zoom = {
-  MIN: 25,
-  MAX: 100,
-  STEP: 25
-};
-
+import { ZOOM } from './constants.js';
 const scaleInput = document.querySelector('.scale__control--value');
 
 const changeScale = (factor = 1) => {
   let newValue = parseInt(scaleInput.value, 10);
-  newValue = newValue + Zoom.STEP * factor;
-  if (newValue > Zoom.MAX) {
-    newValue = Zoom.MAX;
+  newValue = newValue + ZOOM.STEP * factor;
+  if (newValue > ZOOM.MAX) {
+    newValue = ZOOM.MAX;
   }
-  if (newValue < Zoom.MIN) {
-    newValue = Zoom.MIN;
+  if (newValue < ZOOM.MIN) {
+    newValue = ZOOM.MIN;
   }
   scaleInput.value = `${newValue}%`;
   document.querySelector('.img-upload__preview img').style.transform = `scale(${newValue / 100})`;
