+++ js/util.js
@@ -1,4 +1,4 @@
-const TIME_TO_DELETE_MESSAGE = 5000;
+import { TIME_TO_DELETE_MESSAGE } from './constants.js';
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
@@ -10,4 +10,41 @@ const showDataErrorMessage = () => {
   }, TIME_TO_DELETE_MESSAGE);
 };
 
-export {isEscapeKey, showDataErrorMessage};
+const removeMessage = () => {
+  document.body.lastChild.remove();
+  document.removeEventListener('keydown', onEscapeBtnClick);
+};
+
+function onEscapeBtnClick (evt) {
+  if (isEscapeKey(evt)) {
+    removeMessage();
+  }
+}
+
+const renderMessage = (element) => {
+  const message = element.cloneNode(true);
+
+  message.querySelector('button').addEventListener('click', () => {
+    removeMessage();
+  });
+
+  message.addEventListener('click', (evt) => {
+    if (evt.target.classList.contains('success') || evt.target.classList.contains('error')) {
+      removeMessage();
+    }
+  });
+
+  document.addEventListener('keydown', onEscapeBtnClick);
+
+  document.body.append(message);
+};
+
+const showErrorMessage = () => {
+  renderMessage(document.querySelector('#error').content.querySelector('.error'));
+};
+
+const showSuccessMessage = () => {
+  renderMessage(document.querySelector('#success').content.querySelector('.success'));
+};
+
+export { isEscapeKey, showDataErrorMessage, showErrorMessage, showSuccessMessage};
